refactor(storage): rename loadsave.hpp to storage.hpp and update references. Centralize logic into presetManager and leave platform implementation in storage
