+++ firmware-CLK/lib/storage.hpp
@@ -1,77 +1,14 @@
 #pragma once
 
-#include <Arduino.h>
+// storage.hpp — Platform-specific storage backend
+//
+// Only responsibility: persist LoadSaveParams and CalibrationData to
+// non-volatile storage. Schema and business logic live in presetManager.hpp.
+//
+// Swap this file for a new platform (e.g. std::fstream for VCVRack) without
+// touching any other module.
 
-#include "outputs.hpp"
-
-// Number of preset slots.
-// Slot 0 = auto-load/save on boot; slots 1–(NUM_SLOTS-1) = user presets.
-// Increasing this shifts EEPROM_CAL_BASE — re-run CV calibration after changing.
-#define NUM_SLOTS 5
-
-// Struct to hold params that are saved/loaded to/from EEPROM/flash
-// VALID_MAGIC: written to the `valid` field on save; checked on load.
-// Using 0xA5 avoids false-positives from erased flash (0xFF) or zeroed RAM.
-#define VALID_MAGIC 0xA5
-
-struct LoadSaveParams {
-    uint8_t valid;  // VALID_MAGIC (0xA5) = valid data; any other value = use defaults
-    unsigned int BPM;
-    unsigned int externalClockDivIdx;
-    int divIdx[NUM_OUTPUTS];
-    int dutyCycle[NUM_OUTPUTS];
-    bool outputState[NUM_OUTPUTS];
-    uint32_t outputLevel[NUM_OUTPUTS];
-    int outputOffset[NUM_OUTPUTS];
-    int swingIdx[NUM_OUTPUTS];
-    int swingEvery[NUM_OUTPUTS];
-    int pulseProbability[NUM_OUTPUTS];
-    EuclideanParams euclideanParams[NUM_OUTPUTS];
-    int phaseShift[NUM_OUTPUTS];
-    int waveformType[NUM_OUTPUTS];
-    byte CVInputTarget[NUM_CV_INS];
-    int CVInputAttenuation[NUM_CV_INS];
-    int CVInputOffset[NUM_CV_INS];
-    EnvelopeParams envParams[NUM_OUTPUTS];
-    QuantizerParams quantizerParams[NUM_OUTPUTS];
-};
-
-// CV input calibration — stored past the preset block so preset operations
-// never clobber it. Only updated by the calibration routine (Phase 8B).
-struct CalibrationData {
-    boolean valid;
-    float cvCalOffset[NUM_CV_INS];  // back-calculated intercept from 1V anchor
-    float cvCalScale[NUM_CV_INS];   // 1638 / (reading_3v - reading_1v)
-};
-
-// ── Load default setting data ─────────────────────────────────────────
-LoadSaveParams LoadDefaultParams() {
-    LoadSaveParams p;
-    p.valid = VALID_MAGIC;
-    p.BPM = 120;
-    p.externalClockDivIdx = 0;
-    for (int i = 0; i < NUM_OUTPUTS; i++) {
-        p.divIdx[i] = 9;
-        p.dutyCycle[i] = 50;
-        p.outputState[i] = true;
-        p.outputLevel[i] = 100;
-        p.outputOffset[i] = 0;
-        p.swingIdx[i] = 0;
-        p.swingEvery[i] = 2;
-        p.pulseProbability[i] = 100;
-        p.euclideanParams[i] = {false, 10, 6, 1, 0};
-        p.phaseShift[i] = 0;
-        p.waveformType[i] = 0;
-        p.envParams[i] = {200.0f, 200.0f, 70.0f, 250.0f, 0.5f, 0.5f, 0.5f, false};
-        p.quantizerParams[i] = {false, 3, 4, 1, 0};
-    }
-    for (int i = 0; i < NUM_CV_INS; i++) {
-        p.CVInputTarget[i] = 0;
-        p.CVInputAttenuation[i] = 0;
-        p.CVInputOffset[i] = 0;
-    }
-    return p;
-}
+#include "presetManager.hpp"
 
 #if defined(ARDUINO_ARCH_RP2040)
 // ── RP2040: arduino-pico EEPROM emulation ────────────────────────────
